Refactor: Move validateDepositInputs from TradeDataValidation to DepositTxValidation
