+++ core/src/main/java/bisq/core/trade/validation/DepositTxValidation.java
@@ -21,6 +21,8 @@ import bisq.core.btc.model.RawTransactionInput;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
 import bisq.core.offer.Offer;
+import bisq.core.trade.model.bisq_v1.Trade;
+import bisq.core.trade.validation.exceptions.InvalidTxException;
 
 import org.bitcoinj.core.Coin;
 import org.bitcoinj.core.NetworkParameters;
@@ -34,11 +36,14 @@ import com.google.common.annotations.VisibleForTesting;
 import java.util.Arrays;
 import java.util.List;
 
+import lombok.extern.slf4j.Slf4j;
+
 import static bisq.core.trade.validation.TransactionValidation.checkTransaction;
 import static bisq.core.util.Validator.checkNonEmptyBytes;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 
+@Slf4j
 public final class DepositTxValidation {
     public static final double MAX_TRADE_PRICE_DEVIATION = 1.5;
 
@@ -208,5 +213,26 @@ public final class DepositTxValidation {
             inputValue = Math.addExact(inputValue, input.value);
         }
         return inputValue;
+    }
+
+    public static void validateDepositInputs(Trade trade) throws InvalidTxException {
+        // assumption: deposit tx always has 2 inputs, the maker and taker
+        if (trade == null || trade.getDepositTx() == null || trade.getDepositTx().getInputs().size() != 2) {
+            throw new InvalidTxException("Deposit transaction is null or has unexpected input count");
+        }
+        Transaction depositTx = trade.getDepositTx();
+        String txIdInput0 = depositTx.getInput(0).getOutpoint().getHash().toString();
+        String txIdInput1 = depositTx.getInput(1).getOutpoint().getHash().toString();
+        String contractMakerTxId = trade.getContract().getOfferPayload().getOfferFeePaymentTxId();
+        String contractTakerTxId = trade.getContract().getTakerFeeTxID();
+        boolean makerFirstMatch = contractMakerTxId.equalsIgnoreCase(txIdInput0) && contractTakerTxId.equalsIgnoreCase(txIdInput1);
+        boolean takerFirstMatch = contractMakerTxId.equalsIgnoreCase(txIdInput1) && contractTakerTxId.equalsIgnoreCase(txIdInput0);
+        if (!makerFirstMatch && !takerFirstMatch) {
+            String errMsg = "Maker/Taker txId in contract does not match deposit tx input";
+            log.error(errMsg +
+                    "\nContract Maker tx=" + contractMakerTxId + " Contract Taker tx=" + contractTakerTxId +
+                    "\nDeposit Input0=" + txIdInput0 + " Deposit Input1=" + txIdInput1);
+            throw new InvalidTxException(errMsg);
+        }
     }
 }
