+++ desktop/src/main/java/bisq/desktop/main/portfolio/pendingtrades/steps/seller/SellerStep1View.java
@@ -22,7 +22,7 @@ import bisq.desktop.main.portfolio.pendingtrades.PendingTradesViewModel;
 import bisq.desktop.main.portfolio.pendingtrades.steps.TradeStepView;
 
 import bisq.core.locale.Res;
-import bisq.core.trade.bisq_v1.TradeDataValidation;
+import bisq.core.trade.validation.DepositTxValidation;
 import bisq.core.trade.validation.exceptions.ValidationException;
 
 public class SellerStep1View extends TradeStepView {
@@ -81,7 +81,7 @@ public class SellerStep1View extends TradeStepView {
     // Verify that deposit tx inputs are matching the trade fee txs outputs.
     private void validateDepositInputs() {
         try {
-            TradeDataValidation.validateDepositInputs(trade);
+            DepositTxValidation.validateDepositInputs(trade);
         } catch (ValidationException e) {
             if (!model.dataModel.tradeManager.isAllowFaultyDelayedTxs()) {
                 new Popup().warning(Res.get("portfolio.pending.invalidTx", e.getMessage())).show();
